requirements.txt
fastapi==0.95.2
gunicorn==20.1.0
Jinja2==3.1.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
python-socketio==5.8.0
